kernel: grow the EL0 stack to 40KB (both STACK_PAGES) for the client session
The client-side export session (Decision 4) nests one call level deeper than
the path-based client ops - serve -> drain -> handle_client -> handle_rmount
-> session_rmount -> find_or_open -> client_connect, each its own frame -
so its ~1600-byte packet buffers ran ~1-3KB past the 32KB stack the one-shot
path (one level shallower) just fit. Same growth as every prior netd client
op that outgrew the stack (24->32KB before); +8KB, RAM ample.

STACK_PAGES is DUPLICATED in loader.rs and mmu.rs (the loader allocates the
region, mmu.rs places the guard page STACK_PAGES+1 from the end). Both move
to 10 together: changing only one misplaces the guard mid-stack and turns a
clean overflow fault into erratic corruption - which cost real debugging time
this session before the mismatch was spotted. The mmu.rs comment already said
'must match loader.rs'; now it does again.

libc/cbig.c: print byte counts with %d and (int) casts - this libc's printf
has no 'l' length modifier, so %ld printed literally.

With this, cbig reads /man/grep (4661 bytes, nine chunks) over a remote mount
through a held session and it byte-matches the local copy - the client half
of step 5 working end to end on the two-VM rig.

+++ libc/cbig.c
@@ -72,12 +72,12 @@ int main(void) {
     int longer = ln > 512;
     int same = (ln == rn) && (memcmp(local, remote, (size_t)ln) == 0);
     if (same && longer) {
-        printf("cbig: %ld bytes over the remote mount match the local copy "
-               "(%ld > one chunk)\r\n", rn, ln);
+        printf("cbig: %d bytes over the remote mount match the local copy "
+               "(%d > one chunk)\r\n", (int)rn, (int)ln);
         return 0;
     }
-    printf("cbig: MISMATCH - local %ld bytes, remote %ld bytes, %s, %s\r\n",
-           ln, rn, same ? "equal" : "DIFFER",
+    printf("cbig: MISMATCH - local %d bytes, remote %d bytes, %s, %s\r\n",
+           (int)ln, (int)rn, same ? "equal" : "DIFFER",
            longer ? "longer than a chunk" : "NOT longer than a chunk");
     return 1;
 }
